internals: splitting up decls/defs for quadrature classes
towards fixing #84
