+++ include/pfasst/quadrature/clenshaw_curtis.hpp
@@ -30,41 +30,24 @@ namespace pfasst
 
       public:
         //! @{
-        explicit ClenshawCurtis(const size_t num_nodes)
-          : IQuadrature<precision>(num_nodes)
-        {
-          if (this->num_nodes < 2) {
-            throw invalid_argument("Clenshaw-Curtis quadrature requires at least two quadrature nodes.");
-          }
-          this->compute_nodes();
-          this->compute_weights();
-        }
-
+        explicit ClenshawCurtis(const size_t num_nodes);
         ClenshawCurtis() = default;
-
         virtual ~ClenshawCurtis() = default;
         //! @}
 
         //! @{
-        virtual bool left_is_node() const { return LEFT_IS_NODE; }
-
-        virtual bool right_is_node() const { return RIGHT_IS_NODE; }
+        virtual bool left_is_node() const override;
+        virtual bool right_is_node() const override;
         //! @}
 
       protected:
         //! @{
-        virtual void compute_nodes()
-        {
-          this->nodes = vector<precision>(this->num_nodes, precision(0.0));
-          auto roots = Polynomial<precision>::legendre(this->num_nodes).roots();
-
-          for (size_t j = 0; j < this->num_nodes; j++) {
-            this->nodes[j] = 0.5 * (1.0 - cos(j * pi<precision>() / (this->num_nodes - 1)));
-          }
-        }
+        virtual void compute_nodes() override;
         //! @}
     };
   }  // ::pfasst::quadrature
 }  // ::pfasst
+
+#include "clenshaw_curtis_impl.hpp"
 
 #endif  // _PFASST__QUADRATURE__CLENSHAW_CURTIS_HPP_
