+++ include/pfasst/quadrature/gauss_legendre.hpp
@@ -1,14 +1,7 @@
 #ifndef _PFASST__QUADRATURE__GAUSS_LEGENDRE_HPP_
 #define _PFASST__QUADRATURE__GAUSS_LEGENDRE_HPP_
 
-#include <cassert>
-#include <vector>
-
-#include "../interfaces.hpp"
-#include "polynomial.hpp"
 #include "interface.hpp"
-
-using namespace std;
 
 
 namespace pfasst
@@ -27,38 +20,24 @@ namespace pfasst
 
       public:
         //! @{
-        explicit GaussLegendre(const size_t num_nodes)
-          : IQuadrature<precision>(num_nodes)
-        {
-          this->compute_nodes();
-          this->compute_weights();
-        }
-
+        explicit GaussLegendre(const size_t num_nodes);
         GaussLegendre() = default;
-
         virtual ~GaussLegendre() = default;
         //! @}
 
         //! @{
-        virtual bool left_is_node() const { return LEFT_IS_NODE; }
-
-        virtual bool right_is_node() const { return RIGHT_IS_NODE; }
+        virtual bool left_is_node() const override;
+        virtual bool right_is_node() const override;
         //! @}
 
       protected:
         //! @{
-        virtual void compute_nodes() override
-        {
-          this->nodes = vector<precision>(this->num_nodes, precision(0.0));
-          auto roots = Polynomial<precision>::legendre(this->num_nodes).roots();
-
-          for (size_t j = 0; j < this->num_nodes; j++) {
-            this->nodes[j] = 0.5 * (1.0 + roots[j]);
-          }
-        }
+        virtual void compute_nodes() override;
         //! @}
     };
   }  // ::pfasst::quadrature
 }  // ::pfasst
+
+#include "gauss_legendre_impl.hpp"
 
 #endif  // _PFASST__QUADRATURE__GAUSS_LEGENDRE_HPP_
