+++ include/pfasst/quadrature/gauss_lobatto.hpp
@@ -1,14 +1,7 @@
 #ifndef _PFASST__QUADRATURE__GAUSS_LOBATTO_HPP_
 #define _PFASST__QUADRATURE__GAUSS_LOBATTO_HPP_
 
-#include <cassert>
-#include <vector>
-
-#include "../interfaces.hpp"
-#include "polynomial.hpp"
 #include "interface.hpp"
-
-using namespace std;
 
 
 namespace pfasst
@@ -27,43 +20,24 @@ namespace pfasst
 
       public:
         //! @{
-        explicit GaussLobatto(const size_t num_nodes)
-          : IQuadrature<precision>(num_nodes)
-        {
-          if (this->num_nodes < 2) {
-            throw invalid_argument("Gauss-Lobatto quadrature requires at least two quadrature nodes.");
-          }
-          this->compute_nodes();
-          this->compute_weights();
-        }
-
+        explicit GaussLobatto(const size_t num_nodes);
         GaussLobatto() = default;
-
         virtual ~GaussLobatto() = default;
         //! @}
 
         //! @{
-        virtual bool left_is_node() const { return LEFT_IS_NODE; }
-
-        virtual bool right_is_node() const { return RIGHT_IS_NODE; }
+        virtual bool left_is_node() const override;
+        virtual bool right_is_node() const override;
         //! @}
 
       protected:
         //! @{
-        virtual void compute_nodes() override
-        {
-          this->nodes = vector<precision>(this->num_nodes, precision(0.0));
-          auto roots = Polynomial<precision>::legendre(this->num_nodes - 1).differentiate().roots();
-
-          for (size_t j = 0; j < this->num_nodes - 2; j++) {
-            this->nodes[j + 1] = 0.5 * (1.0 + roots[j]);
-          }
-          this->nodes.front() = 0.0;
-          this->nodes.back() = 1.0;
-        }
+        virtual void compute_nodes() override;
         //! @}
     };
   }  // ::pfasst::quadrature
 }  // ::pfasst
+
+#include "gauss_lobatto_impl.hpp"
 
 #endif  // _PFASST__QUADRATURE__GAUSS_LOBATTO_HPP_
