+++ include/pfasst/quadrature/uniform.hpp
@@ -1,13 +1,7 @@
 #ifndef _PFASST__QUADRATURE__UNIFORM_HPP_
 #define _PFASST__QUADRATURE__UNIFORM_HPP_
 
-#include <cassert>
-#include <vector>
-
-#include "../interfaces.hpp"
 #include "interface.hpp"
-
-using namespace std;
 
 
 namespace pfasst
@@ -26,39 +20,24 @@ namespace pfasst
 
       public:
         //! @{
-        explicit Uniform(const size_t num_nodes)
-          : IQuadrature<precision>(num_nodes)
-        {
-          if (this->num_nodes < 2) {
-            throw invalid_argument("Uniform quadrature requires at least two quadrature nodes.");
-          }
-          this->compute_nodes();
-          this->compute_weights();
-        }
-
+        explicit Uniform(const size_t num_nodes);
         Uniform() = default;
-
         virtual ~Uniform() = default;
         //! @}
 
         //! @{
-        virtual bool left_is_node() const { return LEFT_IS_NODE; }
-
-        virtual bool right_is_node() const { return RIGHT_IS_NODE; }
+        virtual bool left_is_node() const override;
+        virtual bool right_is_node() const override;
         //! @}
 
       protected:
         //! @{
-        virtual void compute_nodes()
-        {
-          this->nodes = vector<precision>(this->num_nodes, precision(0.0));
-          for (size_t j = 0; j < this->num_nodes; j++) {
-            this->nodes[j] = precision(j) / (this->num_nodes - 1);
-          }
-        }
+        virtual void compute_nodes() override;
         //! @}
     };
   }  // ::pfasst::quadrature
 }  // ::pfasst
+
+#include "uniform_impl.hpp"
 
 #endif  // _PFASST__QUADRATURE__UNIFORM_HPP_
